Biomass: fetch on the 20th, and survive a null Utsettsar (#14)
* Run biomass fetch on the 20th, not the 21st

Fiskeridirektoratet publishes the monthly biomass file on the 20th
(Last-Modified 2026-09-20T04:38Z this month), but the cron fired on the
21st, leaving BigQuery a month stale for a full day every month.

Move the schedule to the 20th at 06:00 UTC and keep the 21st as a
backstop in case FD publishes late. fetch_biomass.py is a guarded
truncate-and-reload (aborts on 0 rows), so the extra run is a no-op.


* Don't crash when Utsettsar is null

The 2026-08 file carries one stray PO 1 row — a single fish, 1 g, no
year-class — and `(Ar - Utsettsar).astype(int)` cannot hold NA, so
clean() raised IntCastingNaNError and the whole load aborted.

Compute running_month as nullable Int64 instead. The row is kept with a
null running_month rather than dropped: it is FD's data, it moves no
totals (0.001 kg), and cohort logic already ignores null year-classes.
The BigQuery column is NULLABLE.

Verified against the live feed: 5,246 rows clean, 2026-08 present,
941,145,751 kg standing biomass, one null running_month.


---------

+++ fetch_biomass.py
@@ -76,7 +76,11 @@ def clean(df: pd.DataFrame) -> pd.DataFrame:
 
     # Add AarMnd computed column
     df["AarMnd"] = df["Ar"].astype(str) + "-" + df["Maaned_kode"].astype(str).str.zfill(2)
-    df["running_month"] = (df["Ar"] - df["Utsettsar"]).astype(int) * 12 + df["Maaned_kode"].astype(int)
+    # Utsettsar is null on the odd stray row (e.g. PO 1, 2026-08: one fish, 1 g),
+    # so keep this nullable rather than casting straight to int.
+    df["running_month"] = (
+        (df["Ar"] - df["Utsettsar"]) * 12 + df["Maaned_kode"]
+    ).astype("Int64")
 
     missing = [c for c in KEEP_COLS if c not in df.columns]
     if missing:
